Extract Partitioner and PotentialMergedWork from MergerRule
