+++ pipeline/merger/src/main/scala/uk/ac/wellcome/platform/merger/rules/MergerRule.scala
@@ -4,7 +4,7 @@ import uk.ac.wellcome.models.work.internal.{
   TransformedBaseWork,
   UnidentifiedWork
 }
-import uk.ac.wellcome.platform.merger.model.MergedWork
+import uk.ac.wellcome.platform.merger.model.{MergedWork, PotentialMergedWork}
 
 trait MergerRule { this: Partitioner with WorkPairMerger =>
 
@@ -26,41 +26,6 @@ trait MergerRule { this: Partitioner with WorkPairMerger =>
           redirectedWork
         )
     }
-}
-
-case class PotentialMergedWork(target: UnidentifiedWork,
-                               redirectedWork: TransformedBaseWork)
-
-case class Partition(potentialMergedWork: PotentialMergedWork,
-                     remainingWorks: Seq[BaseWork])
-
-trait Partitioner {
-  protected def partitionWorks(works: Seq[BaseWork]): Option[Partition]
-}
-
-trait WorkTagPartitioner extends Partitioner {
-
-  sealed trait WorkTag
-
-  case object Target extends WorkTag
-  case object Redirected extends WorkTag
-  case object PassThrough extends WorkTag
-
-  protected def tagWork(work: BaseWork): WorkTag
-
-  def partitionWorks(works: Seq[BaseWork]): Option[Partition] = {
-    val taggedWorks = works.groupBy(tagWork)
-    val targetWorks = taggedWorks.get(Target).toList.flatten
-    val redirectedWorks = taggedWorks.get(Redirected).toList.flatten
-    val remaining = taggedWorks.get(PassThrough).toList.flatten
-    (targetWorks, redirectedWorks) match {
-      case (
-          List(target: UnidentifiedWork),
-          List(redirected: TransformedBaseWork)) =>
-        Some(Partition(PotentialMergedWork(target, redirected), remaining))
-      case _ => None
-    }
-  }
 }
 
 trait WorkPairMerger {
